requirements.txt
beautifulsoup4==4.12.3
pypdf==5.1.0
python-dateutil==2.9.0.post0
requests==2.32.3

# Optional: hybrid embedding tagging
sentence-transformers==3.1.1
scikit-learn==1.5.2

# Optional: LLM-assisted classification
openai==1.51.2
